OpenMP Offloading is mostly working. I think it was because gamin is declare on the stack globally, there was an issue with copying it.
Few issues:
-Reading in, Poly was wrong. Everything else looked okay
-It is slower than a single core. That could be poor memory management. Declaring things as target and updating them only when necessary
	can eliminate that overhead significantly. Especially for the trial fields

Few more issues. Diquark and PBP are returning 0 rather than almost zero

+++ INCLUDE/su2hmc.h
@@ -51,7 +51,7 @@ extern Complex_f gamval_f[5][4];
 __managed__ int *gamin_d;
 __managed__ 
 #endif 
-extern short 
+extern int 
 #ifndef __NVCC__ 
 __attribute__((aligned(AVX)))
 #endif
